Irrigation_Main_App/requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
